pyproject.toml
[project]
name = "resume_crew"
version = "0.1.0"
description = "Resume optimization crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.95.0,<1.0.0",
    "pydantic>=2.0.0",
    "crewai-tools>=0.12.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
resume_crew = "resume_crew.main:run"
run_crew = "resume_crew.main:run"
train = "resume_crew.main:train"
replay = "resume_crew.main:replay"
test = "resume_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
